Rename cloud.CloudTypeCAAS to cloud.CloudTypeKubernetes;
